Added env file and started working on getting the evendemy service to simply serve a main.js file with nodejs
Got a version, that runs both postgresql and evendemy
with docker-compose. This is working.
Now i need to add the volume for pictures.

Added volumes for user and meeting images

Changed Dockerfile for backend to have seperate build and run.
docker-compose.yaml no longer has a command for running
the built evendemy docker.

Fixed a "typo" in the Dockerfile in the backend

Changed docker run stage for backend to slim.

Added a service user with a home direcory and removed
its write and execute privileges for everything except the image assets

Removed unused database from docker-compose.
Now it's just here so i can pass in the thing from .evn easily

Somehow the volumes for the images dont really work
so now i just added a user and i switch to it for running the backend

Changed some volume things

Using the users home directory for running stopped working.
Now it is run from /usr/src/backend

Started working on a UI Middleware Component,
that should intercept everything with the prefix
/ui and send it to the webapp

Fixed a typo

Can render the webapp now, but the js and css files won't load.

Done with serving the frontend from the backend. No need for UI middleware anymore.
The UI is served through the webapps dist folder.

Forgot to add nestjs/serve-static in the other pr.

With host networking enabled we can work with a database
outside of docker compose.

Got serving the frontend working in the docker container.
This is now done i think

Fixed and tested transloco localization service to
use the correct path when running from the backend
in /ui

Removed 'ls' that i used for debugging.

Trying to get CORS to work

Frontend is rendered / and the api is behind /api now.
Set backend url to include /api now

Changed old backend in server/ to also use /api as
its api prefix. Thank god theres vim for replacing this fast.
Changed backend_url in webapp config to localhost:8080/api.
Added api_interceptor to webapp so we have cors on all localhost ports

Changed everything mentioned in the comments
of the PR

Changed image paths