Added realm option to auth.module.ts, so we now use the right realm with keycloak. localhost:8080/api/auth is now unprotected to allow logins. I think I need to create users in keycloak now for testing
Redid the backend config of Keycloak. Added
a config option for the keycloak secrets. Started working on getting
webapp to use keycloak as login, but i dont know which uri to redirect to.

Login is not needed anymore. It's still there, but
i think we dont need it anymore. This now redirects to our keycloak login.

Redirect URI still doesnt redirect right and
configservice says it isn't defined in the webapp. But we now have the
Keycloak configuration things in env and as and env var in docker, so
we dont have to hardcode them anymore

Redirecting from our keycloak auth now works, but
the application crashes, as soon, as it tries to load the config service

adding validityPeriode for ideas

Docker, Static Serve, api on /api in new and old backend

Did the required changes to upgrade from angular 13 to 14

Upgraded from angular 14 to 15. Needed to change
some css imports, but as of now this is working with angular 15

Upgraded from angular 15 to 16. Angular 17 is available,
but 16 is more than new enough

Added hammerjs for ngx-image-cropper

Added hammerjs

Added hammer js

Fixed dockerfile, so that we can now write to the image volumes

Fixed dialog hammerjs for ngx-image-cropper,
editable-text and meeting image stuff

Changed the way images are retrieved.
This should resolve all issues in the PR

Did the required changes to upgrade from angular 13 to 14

Upgraded from angular 14 to 15. Needed to change
some css imports, but as of now this is working with angular 15

Dockerfile is now better. I can't describe it differently

This is version of docker now has chowns for copy

Reverted dockerfile to running version

Fixed needing to log in to view images from the api
with pipes.

Fixed the reload bug with the config service.
Fixed no username in user-image-component.

Removed every last trace of getLoggedInUsername.
Changed LoggedInGuardService to work with keycloak

Everything is undefined after the redirect. Website is not loading.
I dont know why. Tip: Look for circular deps. Sorry this is not working :(

I did not upload a secret to github